+++ docs/release-notes.rst
@@ -26,6 +26,7 @@ Maven Nexus repository. The server is deployed as a Docker image to a Docker reg
 Version 1.14.0, 19 February 2018
 -------------------------------
 
+* Add search-by-date method for federation subscription update (ACUMOS-61)
 * Add peer group, solution group and mapping features for federation access control (ACUMOS-62)
 * Add feature for User notification preference and user notification (ACUMOS-166) 
 * Refactor to drop code-name database tables (ACUMOS-144)